refactor: AttemptCount 클래스를 entity 패키지로 이동
