+++ src/main/kotlin/racingcar/entity/AttemptCount.kt
@@ -1,4 +1,4 @@
-package racingcar.model
+package racingcar.entity
 
 import racingcar.misc.Values
